+++ frontend/src/pages/config/entities/editors/fixture-editor.tsx
@@ -8,6 +8,7 @@ import {
   FormTextInput,
   FormNumberInput,
   FormEntityReferenceSelect,
+  FormCheckbox,
 } from '../../../../ui/forms/form-input'
 import { EntityEditorProps } from '../types'
 import { Label } from '../../../../ui/forms/label'
@@ -27,6 +28,12 @@ import {
   getFractionWithMargin,
   getTouchEventOffset,
 } from '../../../../util/touch'
+import { useEvent } from '../../../../hooks/performance'
+
+const checkboxLink = css`
+  display: block;
+  margin-top: ${baseline(2)};
+`
 
 const errorMessage = css`
   margin-top: ${baseline(4)};
@@ -85,6 +92,9 @@ export function FixtureEditor({
   const { masterData, masterDataMaps } = useMasterDataAndMaps()
   const [sharing, setSharing] = useState(!!entry.fixturesSharingChannel)
   const touchRef = useRef<HTMLDivElement>(null)
+  const toggleCreateGroup = useEvent(() =>
+    formState.changeValue('createGroup', !formState.values.createGroup)
+  )
 
   const {
     type,
@@ -264,6 +274,12 @@ export function FixtureEditor({
                   </>
                 }
               />
+            )}
+            {!id && (
+              <a className={checkboxLink} onClick={toggleCreateGroup}>
+                <FormCheckbox formState={formState} name="createGroup" inline />{' '}
+                Create group containing the fixture
+              </a>
             )}
             {overlappingChannels.length > 0 && (
               <div className={errorMessage}>